changed from int to float to accomodate it being a price

+++ bangazonapi/views/product.py
@@ -275,7 +275,7 @@ class Products(ViewSet):
         
         if min_price is not None:
             def price_filter(product):
-                if product.price >= int(min_price):
+                if product.price >= float(min_price):
                     return True
                 return False
